+++ lib/router.ex
@@ -16,6 +16,10 @@ defmodule Tamnoon.Router do
     send_file(conn, 200, "tamnoon_out/app.html")
   end
 
+  get "/ws_connect.js" do
+    send_file(conn, 200, Application.app_dir(:tamnoon, "priv/static/ws_connect.js"))
+  end
+
   match _ do
     send_resp(conn, 404, "404")
   end